change amount feilds to be float

+++ budget/models.py
@@ -14,7 +14,7 @@ class BudgetEntry(models.Model):
     description = models.CharField(max_length=200)
     pub_date = models.DateTimeField('date start')
     end_date = models.DateTimeField('date end')
-    amount = models.PositiveIntegerField(default=0)
+    amount = models.FloatField(default=0)
     budget_cycle = models.PositiveSmallIntegerField(choices=PAYMENT_CYCLE, default=1)
     def __str__(self):
             return self.description
@@ -39,7 +39,7 @@ class Expense(models.Model):
     entry = models.ForeignKey(BudgetEntry, on_delete=models.CASCADE)
     description = models.CharField(max_length=200)
     expense_date = models.DateTimeField('date published')
-    amount = models.PositiveIntegerField(default=0)
+    amount = models.FloatField(default=0)
     payment_method = models.PositiveSmallIntegerField(choices=PAYMENT, default=2)
     payments = models.PositiveSmallIntegerField(default=1)
     user = models.CharField(max_length=200, default='ariel')
